refactor: login, plan list screen
