+++ package.json
@@ -42,7 +42,8 @@
     "react-native-web": "0.17.1",
     "recoil": "^0.6.1",
     "styled-components": "^5.3.3",
-    "@react-native-community/datetimepicker": "4.0.0"
+    "@react-native-community/datetimepicker": "4.0.0",
+    "expo-image-picker": "~12.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
