+++ src/constants/message.js
@@ -1,0 +1,5 @@
+const MESSAGE = {
+  ERROR: "Error",
+};
+
+export default MESSAGE;
