+++ src/screens/LoginScreen.js
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from "react";
+import { StyleSheet, Button, View, Text } from "react-native";
 import { useSetRecoilState } from "recoil";
 import asyncStorage from "@react-native-async-storage/async-storage";
 import * as WebBrowser from "expo-web-browser";
@@ -9,11 +10,10 @@ import {
   GoogleAuthProvider,
   signInWithCredential,
 } from "firebase/auth";
-import { StyleSheet, Button, View, Text } from "react-native";
 import PropTypes from "prop-types";
 
 import getEnvVars from "../../environment";
-import { getLogin, getUserInfo } from "../../util/api/user";
+import { getLoginApi } from "../../util/api/user";
 import { firebaseConfig } from "../config/firebaseConfig";
 import { userState } from "../states/userState";
 import axios from "../config/axiosConfig";
@@ -26,35 +26,7 @@ WebBrowser.maybeCompleteAuthSession();
 
 export default function LoginScreen({ navigation }) {
   const [firebaseToken, setFirebaseToken] = useState("");
-  const setUserState = useSetRecoilState(userState);
-
-  const checkAccessToken = async () => {
-    try {
-      const accessToken = await asyncStorage.getItem("accessToken");
-      const userId = await asyncStorage.getItem("userId");
-
-      if (!userId) return;
-
-      if (userId) {
-        const userInfo = await getUserInfo(userId);
-        axios.defaults.headers.Authorization = `Bearer ${accessToken}`;
-
-        setUserState({
-          email: userInfo.email,
-          name: userInfo.name,
-          userId: userId,
-        });
-        navigation.navigate("PlanList");
-      }
-    } catch (err) {
-      alert("error");
-      asyncStorage.clear();
-    }
-  };
-
-  useEffect(() => {
-    checkAccessToken();
-  }, []);
+  const setUser = useSetRecoilState(userState);
 
   const [request, response, promptAsync] = Google.useIdTokenAuthRequest({
     clientId: REACT_NATIVE_ANDROID_CLIENT_ID,
@@ -71,15 +43,15 @@ export default function LoginScreen({ navigation }) {
   }, [response]);
 
   useEffect(() => {
-    const getLoginApi = async () => {
-      const userInfo = await getLogin(firebaseToken);
+    const getLogin = async () => {
+      const userInfo = await getLoginApi(firebaseToken);
 
       await asyncStorage.setItem("accessToken", userInfo.accessToken);
       await asyncStorage.setItem("userId", userInfo.userId);
 
       axios.defaults.headers.Authorization = `Bearer ${userInfo.accessToken}`;
 
-      setUserState({
+      setUser({
         email: userInfo.email,
         name: userInfo.name,
         userId: userInfo.userId,
@@ -88,7 +60,7 @@ export default function LoginScreen({ navigation }) {
       navigation.navigate("PlanList");
     };
 
-    getLoginApi();
+    getLogin();
   }, [firebaseToken]);
 
   return (
@@ -111,11 +83,12 @@ const styles = StyleSheet.create({
     flex: 1,
     backgroundColor: "#d3edf7",
     alignItems: "center",
-    justifyContent: "space-evenly",
+    justifyContent: "center",
   },
   loginText: {
     fontSize: 60,
     color: "#0a80ae",
+    marginBottom: "40%",
   },
   loginButton: {
     width: 80,
