+++ util/api/user.js
@@ -1,6 +1,6 @@
 import axios from "../../src/config/axiosConfig";
 
-export const getLogin = async (token) => {
+export const getLoginApi = async (token) => {
   const response = await axios.post("/login", {
     headers: {
       Authorization: `Bearer ${token}`,
@@ -10,7 +10,7 @@ export const getLogin = async (token) => {
   return response.data.data;
 };
 
-export const getUserInfo = async (userId) => {
+export const getUserInfoApi = async (userId) => {
   const response = await axios.get(`/user/${userId}`);
 
   return response.data.data;
